# .scratch/mozhou-mvp/contracts/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "墨舟 MoZhou API",
    "version": "1.0.0",
    "description": "AI 小说写作平台 — 全站 19 组端点契约（UVSD Contract-Driven）"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署"
    }
  ],
  "paths": {
...
    "/api/v1/novels": {
      "get": {
        "summary": "作品列表",
        "operationId": "novelList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "novels": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建作品",
        "operationId": "novelCreate",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "novel": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/novels/{id}/chapters": {
      "post": {
        "summary": "新建章节",
        "operationId": "chapterCreate",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          }
        }
      },
      "patch": {
        "summary": "更新章节（标题/状态/正文；正文保存必须携带 expectedRevision 做乐观并发）",
        "operationId": "chapterUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "chapterId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "status": { "type": "string", "enum": ["draft", "final"] },
                  "content": { "type": "string" },
                  "expectedRevision": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "比较并交换：与当前 revision 不一致 → 409 ContentChanged；未携带则无条件写入（客户端保存正文时必须携带）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的章节行（含 revision）",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "400": {
            "description": "参数缺失/非法",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "404": {
            "description": "章节不存在或不归属当前用户",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "409": {
            "description": "ContentChanged：expectedRevision 与当前 revision 不一致；响应含当前章节行供客户端对账",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
    "/api/v1/shelf": {
      "get": {
        "summary": "书架列表",
        "operationId": "shelfList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "books": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "导入书架",
        "operationId": "shelfImport",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "source": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  },
                  "site": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rankings": {
      "get": {
        "summary": "网文扫榜",
        "operationId": "rankings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "boards": {
                      "type": "array"
                    },
                    "rows": {
                      "type": "array"
                    },
                    "degraded": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "board",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/rankings/trends": {
      "get": {
        "summary": "查询单榜两次快照趋势",
        "operationId": "getRankingTrends",
        "parameters": [
          {"name": "board", "in": "query", "required": false, "schema": { "type": "string, 默认 long-hot" }}
        ],
        "responses": {
          "200": {
            "description": "{board, baseline, current, rows, risers, newEntries, gone, degraded}"
          },
          "400": {
            "description": "未知榜单 code:'UNKNOWN_BOARD'"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "该榜暂无快照"
          }
        }
      }
    },
...
    "/api/v1/runtime/jobs/{jobId}/events": {
      "get": {
        "summary": "事件游标补发（JSON 轮询，非 SSE）",
        "operationId": "getRuntimeJobEvents",
        "parameters": [
          {"name": "jobId", "in": "path", "required": true, "schema": { "type": "string" }},
          {"name": "afterSeq", "in": "query", "required": false, "schema": { "type": "integer" }}
        ],
        "responses": {
          "200": {
            "description": "seq > afterSeq 的事件（语义对应 Last-Event-ID 续传）"
          },
          "400": {
            "description": "afterSeq 非法"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "任务不存在或不属于当前用户"
          }
        }
      }
    },
...
    "/api/v1/runtime/generations/{generationId}": {
      "get": {
        "summary": "本次创作链路证据",
        "operationId": "getGenerationEvidence",
        "parameters": [
          {"name": "generationId", "in": "path", "required": true, "schema": { "type": "string" }}
        ],
        "responses": {
          "200": {
            "description": "脱敏证据视图 {evidence:{plan, runs, manifest}}"
          },
          "400": {
            "description": "缺少 generationId"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "生成记录不存在或不属于当前用户"
          }
        }
      }
    },
...
  },
...
}